+++ src/logging/IChannel.ts
@@ -1,19 +1,18 @@
 import { LogLevel } from "./LogLevel";
-import { ILogEntry } from "./ILogEntry";
+import { ILogEntry, ILogEntryData } from "./ILogEntry";
 import { IWriter } from "./IWriter";
-import { IStatusEntry } from "./StatusLog/IStatusEntry";
 import { Emitter } from "@src/core/events";
 
 /** An interface describing a single log writing implementation. */
-export interface ILogEntryWriter<TEntry extends ILogEntry = IStatusEntry> {
+export interface ILogEntryWriter<TEntry extends ILogEntry<TData>, TData extends ILogEntryData, TWriterConfig> {
 
     /** Gets the underlying writer implementation. */
-    readonly writer: IWriter<TEntry>;
+    readonly writer: IWriter<TEntry, TData, TWriterConfig>;
 }
 
 /** Interface describing a single target for log-writing output. */
-export interface IChannel<TEntry extends ILogEntry = IStatusEntry>
-    extends ILogEntryWriter<TEntry> {
+export interface IChannel<TEntry extends ILogEntry<TData>, TData extends ILogEntryData, TWriterConfig>
+    extends ILogEntryWriter<TEntry, TData, TWriterConfig> {
 
     /** Gets or sets whether this channel is enabled. */
     enabled: boolean;
@@ -28,7 +27,7 @@ export interface IChannel<TEntry extends ILogEntry = IStatusEntry>
     // pattern: string;
 
     /** The event which fires when the channel is reconfigured. */
-    reconfigured: Emitter<IChannel<TEntry>>;
+    reconfigured: Emitter<IChannel<TEntry, TData, TWriterConfig>>;
 
     /** Checks if the channel is enabled for the specified `level`.
     * @param {LogLevel} level The level to check.