+++ src/logging/ILogBuilder.ts
@@ -3,16 +3,20 @@ import { IChannel } from "./IChannel";
 import { ILog } from "./ILog";
 import { IWriter } from "./IWriter";
 import { LogLevel } from "./LogLevel";
-import { ILogEntry } from "./ILogEntry";
+import { ILogEntry, ILogEntryData } from "./ILogEntry";
 
-export interface ILogBuilder<TEntry extends ILogEntry, TLog extends ILog<TLog, TEntry>> {
+export interface ILogBuilder<
+    TLog extends ILog<TLog, TEntry, TData, TWriterConfig>,
+    TEntry extends ILogEntry<TData>,
+    TData extends ILogEntryData,
+    TWriterConfig> {
     readonly logCreated: Emitter<TLog>;
-    readonly channels: IChannel[];
+    readonly channels: IChannel<TEntry, TData, TWriterConfig>[];
     readonly name: string;
     andGetLogger(): TLog;
-    atLevel(level: LogLevel): ILogBuilder<TEntry, TLog>;
-    newLog(): ILogBuilder<TEntry, TLog>;
-    newLog(logName: string): ILogBuilder<TEntry, TLog>;
-    newChannel(name: string, writer: IWriter<TEntry>): ILogBuilder<TEntry, TLog>;
-    newChannel(name: string, writer: IWriter<TEntry>, level: LogLevel): ILogBuilder<TEntry, TLog>;
+    atLevel(level: LogLevel): ILogBuilder<TLog, TEntry, TData, TWriterConfig>;
+    newLog(): ILogBuilder<TLog, TEntry, TData, TWriterConfig>;
+    newLog(logName: string): ILogBuilder<TLog, TEntry, TData, TWriterConfig>;
+    newChannel(name: string, writer: IWriter<TEntry, TData, TWriterConfig>): ILogBuilder<TLog, TEntry, TData, TWriterConfig>;
+    newChannel(name: string, writer: IWriter<TEntry, TData, TWriterConfig>, level: LogLevel): ILogBuilder<TLog, TEntry, TData, TWriterConfig>;
 }