+++ src/logging/IWriter.ts
@@ -1,12 +1,11 @@
-import { ILogEntry } from "./ILogEntry";
-import { IStatusEntry } from "./StatusLog/IStatusEntry";
+import { ILogEntry, ILogEntryData } from "./ILogEntry";
 
 export interface IPatternWriterConfig {
     pattern: string;
 }
 
 /** An interface describing a log-entry writing implementation. */
-export interface IWriter<TEntry extends ILogEntry = IStatusEntry, TConfig = IPatternWriterConfig> {
+export interface IWriter<TEntry extends ILogEntry<TData>, TData extends ILogEntryData, TConfig> {
 
     /** Writes a log entry.
      * @param {TEntry} entry The log entry to write