+++ src/logging/StatusLog/LogBuilder.ts
@@ -3,31 +3,32 @@ import { throwIfNullOrEmpty, throwIfNullOrUndefined } from "@src/functions/guard
 
 import { IChannel } from "../IChannel";
 import { ILog } from "../ILog";
-import { Channel } from "../Channel";
 import { IWriter } from "../IWriter";
 import { LogLevel } from "../LogLevel";
-import { IStatusEntry } from "./IStatusEntry";
-import { ILogEntry } from "../ILogEntry";
-import { IStatusLog } from "./StatusLog";
+import { ILogEntry, ILogEntryData } from "../ILogEntry";
 import { ILogBuilder } from "../ILogBuilder";
+import { Channel } from "./Channel";
 
 export abstract class LogBuilder<
-    TEntry extends ILogEntry = IStatusEntry,
-    TLog extends ILog<TLog, TEntry> = IStatusLog>
-    implements ILogBuilder<TEntry, TLog> {
+    TLog extends ILog<TLog, TEntry, TData, TWriterConfig>,
+    TEntry extends ILogEntry<TData>,
+    TData extends ILogEntryData,
+    TWriterConfig>
+
+    implements ILogBuilder<TLog, TEntry, TData, TWriterConfig> {
 
     public static DEFAULT_NAME = 'main';
 
     public logCreated: Emitter<TLog> = new Emitter();
 
-    protected _channels: IChannel[] = [];
-    public get channels(): IChannel[] { return this._channels; }
+    protected _channels: IChannel<TEntry, TData, TWriterConfig>[] = [];
+    public get channels(): IChannel<TEntry, TData, TWriterConfig>[] { return this._channels; }
 
     protected _logName = '';
     public get name(): string { return this._logName; }
 
     /** Instantiates a new `LogBuilder` to store the configuration for a logger. */
-    protected abstract createBuilder(logName: string): LogBuilder<TEntry, TLog>;
+    protected abstract createBuilder(logName: string): LogBuilder<TLog, TEntry, TData, TWriterConfig>;
 
     /** Instantiates a new logger based on the configuration stored in this builder. */
     protected abstract createLogger(): TLog;
@@ -45,7 +46,7 @@ export abstract class LogBuilder<
     }
 
     /** Generates or augments a builder for a logger channel that outputs at any level at least as high the specified level. */
-    public atLevel(level: LogLevel): LogBuilder<TEntry, TLog> {
+    public atLevel(level: LogLevel): LogBuilder<TLog, TEntry, TData, TWriterConfig> {
         this._channels.forEach(channel => {
             channel.level = level;
         });
@@ -53,17 +54,17 @@ export abstract class LogBuilder<
     }
 
     /** Configures the builder to generate a new logger. */
-    public newLog(): LogBuilder<TEntry, TLog>;
-    public newLog(logName: string): LogBuilder<TEntry, TLog>;
-    public newLog(logName: string = LogBuilder.DEFAULT_NAME): LogBuilder<TEntry, TLog> {
+    public newLog(): LogBuilder<TLog, TEntry, TData, TWriterConfig>;
+    public newLog(logName: string): LogBuilder<TLog, TEntry, TData, TWriterConfig>;
+    public newLog(logName: string = LogBuilder.DEFAULT_NAME): LogBuilder<TLog, TEntry, TData, TWriterConfig> {
         throwIfNullOrUndefined(logName, 'logName');
         return this.createBuilder(logName);
     }
 
     /** Configures the builder to generate a new logger channel. */
-    public newChannel(name: string, writer: IWriter<TEntry>): LogBuilder<TEntry, TLog>;
-    public newChannel(name: string, writer: IWriter<TEntry>, level: LogLevel): LogBuilder<TEntry, TLog>;
-    public newChannel(name: string, writer: IWriter<TEntry>, level?: LogLevel): LogBuilder<TEntry, TLog> {
+    public newChannel(name: string, writer: IWriter<TEntry, TData, TWriterConfig>): LogBuilder<TLog, TEntry, TData, TWriterConfig>;
+    public newChannel(name: string, writer: IWriter<TEntry, TData, TWriterConfig>, level: LogLevel): LogBuilder<TLog, TEntry, TData, TWriterConfig>;
+    public newChannel(name: string, writer: IWriter<TEntry, TData, TWriterConfig>, level?: LogLevel): LogBuilder<TLog, TEntry, TData, TWriterConfig> {
         throwIfNullOrEmpty(name, 'name');
         throwIfNullOrUndefined(writer, 'writer');
         this._channels = this._channels.filter(it => it.name != name);