+++ src/logging/index.ts
@@ -1,18 +1,17 @@
 
 export {
-    TerminalLog as Log, ConsoleLog,
+    TerminalLog as Log,
+    ConsoleLog,
     LogManager,
     CONSOLE, TERMINAL
 } from './DEFAULTS';
 
 // 'barrel' export
 export { IChannel, ILogEntryWriter } from './IChannel';
-export { ILog, ILoggerBase } from './ILog';
+export { ILog } from './ILog';
 export { ILogEntry, ILogEntryData } from "./ILogEntry";
 export { ILogManager } from './ILogManager';
 export { ILogBuilder } from './ILogBuilder';
 export { IWriter } from './IWriter';
 export { LogLevel } from "./LogLevel";
 export { LogStatus } from "./LogStatus";
-// export { LogBuilder } from './LogBuilder';
-// export { LogManagerClass } from './LogManager';